Sounds for "Love Yourself Ending"
Notes:
All the tracks named explosion are explosions
"Engine" and "Truck running wav" are Truck Sounds
"Risk" is the Dramatic Music to be played when the player chooses the option (Go to Leiria (dramatically)
"Portal" - the wormhole to Leiria
"Dark Piano" will be played twice. The first time is when the ending begins. The second time is when Seby is depressed at the MSS
"Epic Travel on Celtic Roads Version" will be played when Seby is inside the package going to Leiria
"Romantic Episode" will be played when the romantic vibes (blush) appear